backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djoser==2.2.2
gunicorn==21.2.0
idna==3.7
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
